Reconfigure remaining integration-tests for buckets
