+++ java-manta-it/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
- Copyright (c) 2015-2017, Joyent, Inc. All rights reserved.
+ Copyright (c) 2015-2019, Joyent, Inc. All rights reserved.
 
  This Source Code Form is subject to the terms of the Mozilla Public
  License, v. 2.0. If a copy of the MPL was not distributed with this
