+++ java-manta-it/src/test/java/com/joyent/manta/client/MantaClientRangeIT.java
@@ -7,12 +7,15 @@
  */
 package com.joyent.manta.client;
 
+import com.joyent.manta.client.helper.IntegrationTestHelper;
 import com.joyent.manta.config.BaseChainedConfigContext;
 import com.joyent.manta.config.ConfigContext;
 import com.joyent.manta.config.EncryptionAuthenticationMode;
 import com.joyent.manta.config.IntegrationTestConfigContext;
 import com.joyent.manta.config.SettableConfigContext;
 import com.joyent.manta.http.MantaHttpHeaders;
+import com.joyent.test.util.MantaAssert;
+import com.joyent.test.util.MantaFunction;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.testng.Assert;
@@ -26,6 +29,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
 import java.util.UUID;
+
+import static com.joyent.manta.exception.MantaErrorCode.RESOURCE_NOT_FOUND_ERROR;
 
 /**
  * Tests for verifying the correct behavior of range operations performed by
@@ -63,11 +68,13 @@ public class MantaClientRangeIT {
 
     private final String testPathPrefix;
 
-    @Parameters({"encryptionCipher"})
-    public MantaClientRangeIT(final @Optional String encryptionCipher) {
+    public MantaClientRangeIT(final @Optional String encryptionCipher,
+                              final @Optional String testType) throws IOException {
 
         // Let TestNG configuration take precedence over environment variables
         SettableConfigContext<BaseChainedConfigContext> config = new IntegrationTestConfigContext(encryptionCipher);
+        final String testName = this.getClass().getSimpleName();
+
         // Range request have to be in optional authentication mode
         if (config.isClientEncryptionEnabled()) {
             config.setEncryptionAuthenticationMode(EncryptionAuthenticationMode.Optional);
@@ -75,18 +82,19 @@ public class MantaClientRangeIT {
 
         mantaClient = new MantaClient(config);
         this.config = config;
-        testPathPrefix = IntegrationTestConfigContext.generateBasePath(config, this.getClass().getSimpleName());
-
+        testPathPrefix = IntegrationTestHelper.setupTestPath(config, mantaClient,
+                testName, testType);
     }
 
     @BeforeClass
-    public void beforeClass() throws IOException {
-        mantaClient.putDirectory(testPathPrefix, true);
+    @Parameters({"testType"})
+    public void beforeClass(final @Optional String testType) throws IOException {
+        IntegrationTestHelper.createTestBucketOrDirectory(mantaClient, testPathPrefix, testType);
     }
 
     @AfterClass
     public void afterClass() throws IOException {
-        IntegrationTestConfigContext.cleanupTestDirectory(mantaClient, testPathPrefix);
+        IntegrationTestHelper.cleanupTestBucketOrDirectory(mantaClient, testPathPrefix);
     }
 
     public final void canGetWithRangeHeader() throws IOException {
@@ -104,6 +112,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     public final void canGetWithComputedRangeHeader() throws IOException {
@@ -120,6 +132,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     public final void canGetWithUnboundedEndRange() throws IOException {
@@ -137,6 +153,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     public final void canGetWithUnboundedStartRange() throws IOException {
@@ -154,6 +174,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     public final void canGetWithEndRangeBeyondObjectSize() throws IOException {
@@ -171,6 +195,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     public final void canGetWithZeroRange() throws IOException {
@@ -188,6 +216,10 @@ public class MantaClientRangeIT {
             String actual = IOUtils.toString(min, Charset.defaultCharset());
             Assert.assertEquals(actual, expected, "Didn't receive correct range value");
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 
     @Test
@@ -218,5 +250,9 @@ public class MantaClientRangeIT {
                 }
             }
         }
+
+        mantaClient.delete(path);
+        MantaAssert.assertResponseFailureStatusCode(404, RESOURCE_NOT_FOUND_ERROR,
+                (MantaFunction<Object>) () -> mantaClient.head(path));
     }
 }
